// src/app/models/empresa.ts
export interface Empresa {
    Id_Empresa: number,
    razonSocial: string;
    numeroContrato: string;
    rfc: string;
    email: string;
    telefono: string;
    calle: string;
    colonia: string;
    municipio: string;
    estado: string;
    codigo_postal: string;
    representante: string;
}